ask/requirements.txt
Django==1.8.2
Pygments==2.0.2
djangorestframework==3.1.3
httpie==0.9.2
requests==2.7.0
wsgiref==0.1.2
